Crate tlsn_prover

The prover library

This library provides the Prover type. It can be used for creating TLS connections with a server which can be notarized with the help of a notary.

Structs

Closed

State after the TLS connection has been closed.

Initialized

Entry state

Notarize

The state for the Prover during notarization

Prover

A prover instance.

ProverConfig

Configuration for the prover

ProverFuture

Prover future which must be polled for the TLS connection to make progress.

Setup

State after MPC setup has completed.

Enums

ProverError

An error that can occur during proving.

Traits

ProverState