Module tlsn_prover::http

HTTP Prover.

An HTTP prover can be created from a TLS Prover, after the TLS connection has been closed, by calling the to_http method.

The HttpProver provides higher-level APIs for committing and proving data communicated during an HTTP connection.

Modules

state

HTTP prover state.

Structs

HttpCommitmentBuilder

Builder for commitments to data in an HTTP connection.

HttpProofBuilder

Builder for proofs of data in an HTTP connection.

HttpProver

An HTTP prover.

HttpRequestCommitmentBuilder

Builder for commitments to an HTTP request.

HttpResponseCommitmentBuilder

Builder for commitments to an HTTP response.

JsonBody

A JSON body

JsonCommitmentBuilder

Builder for commitments to a JSON value.

JsonProofBuilder

Builder for proofs of a JSON value.

NotarizedHttpSession

A notarized HTTP session.

Enums

HttpCommitmentBuilderError

HTTP commitment builder error.

HttpProofBuilderError

HTTP proof builder error.

HttpProverError

HTTP prover error.

JsonCommitmentBuilderError

JSON commitment builder error.

JsonProofBuilderError

JSON proof builder error.