Module tlsn_prover::tls

TLS prover.

This module provides the TLS prover, which is used with a TLS verifier to prove a transcript of communications with a server.

The TLS prover provides a low-level API, see the HTTP prover which provides abstractions for working with HTTP sessions.

Modules

state

TLS prover states.

Structs

Prover

A prover instance.

ProverConfig

Configuration for the prover

ProverConfigBuilder

Builder for ProverConfig.

ProverControl

A controller for the prover.

ProverFuture

Prover future which must be polled for the TLS connection to make progress.

Enums

ProverConfigBuilderError

Error type for ProverConfigBuilder

ProverError

An error that can occur during proving.