Struct tlsn_prover::tls::Prover

pub struct Prover<T: ProverState> { /* private fields */ }

A prover instance.

Implementations

impl Prover<Notarize>

pub fn sent_transcript(&self) -> &Transcript

Returns the transcript of the sent data.

pub fn recv_transcript(&self) -> &Transcript

Returns the transcript of the received data.

pub fn commitment_builder(&mut self) -> &mut TranscriptCommitmentBuilder

Returns the transcript commitment builder.

pub async fn finalize(self) -> Result<NotarizedSession, ProverError>

Finalizes the notarization returning a NotarizedSession.

impl Prover<Prove>

pub fn sent_transcript(&self) -> &Transcript

Returns the transcript of the sent requests

pub fn recv_transcript(&self) -> &Transcript

Returns the transcript of the received responses

pub fn reveal( &mut self, ranges: impl Into<RangeSet<usize>>, direction: Direction ) -> Result<(), ProverError>

Reveal certain parts of the transcripts to the verifier

This function allows to collect certain transcript ranges. When Prover::prove is called, these ranges will be opened to the verifier.

Arguments

ranges - The ranges of the transcript to reveal

direction - The direction of the transcript to reveal

pub async fn prove(&mut self) -> Result<(), ProverError>

Prove transcript values

pub async fn finalize(self) -> Result<(), ProverError>

Finalize the proving

impl Prover<Initialized>

pub fn new(config: ProverConfig) -> Self

Creates a new prover.

Arguments

config - The configuration for the prover.

pub async fn setup<S: AsyncWrite + AsyncRead + Send + Unpin + 'static>( self, socket: S ) -> Result<Prover<Setup>, ProverError>

Sets up the prover.

This performs all MPC setup prior to establishing the connection to the application server.

Arguments

socket - The socket to the TLS verifier.

impl Prover<Setup>

pub async fn connect<S: AsyncWrite + AsyncRead + Send + Unpin + 'static>( self, socket: S ) -> Result<(TlsConnection, ProverFuture), ProverError>

Connects to the server using the provided socket.

Returns a handle to the TLS connection, a future which returns the prover once the connection is closed.

Arguments

socket - The socket to the server.

impl Prover<Closed>

pub fn sent_transcript(&self) -> &Transcript

Returns the transcript of the sent requests

pub fn recv_transcript(&self) -> &Transcript

Returns the transcript of the received responses

pub fn to_http(self) -> Result<HttpProver<Closed>, HttpProverError>

Creates an HTTP prover.

pub fn start_notarize(self) -> Prover<Notarize>

Starts notarization of the TLS session.

Used when the TLS verifier is a Notary to transition the prover to the next state where it can generate commitments to the transcript prior to finalization.

pub fn start_prove(self) -> Prover<Prove>

Starts proving the TLS session.

This function transitions the prover into a state where it can prove content of the transcript.